requirements-dev.txt
# Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
pytest-mock>=3.12.0
factory-boy>=3.3.0
faker>=22.0.0
coverage[toml]>=7.4.0

# Code Quality
flake8>=7.0.0
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
pylint>=3.0.0

# Security
bandit>=1.7.6
safety>=3.0.0

# Development Tools
ipython>=8.20.0
ipdb>=0.13.13
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
